# rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a education forum server.",
        "title": "Education Forum API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "romgrishin@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
        "/comments/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "list comments with filtering",
                "summary": "List comments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of post",
                        "name": "postId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "show data like XML",
                        "name": "xml",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "default": {
                        "description": ""
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "update comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update comment",
                "parameters": [
                    {
                        "description": "JSON structure for creating post",
                        "name": "RequestPost",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.comment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "default": {
                        "description": ""
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "create comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create comment",
                "parameters": [
                    {
                        "description": "JSON structure for creating post",
                        "name": "RequestPost",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.createCommentStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "201": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/comments/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get comment by ID",
                "summary": "Show comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of comment",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "show data like XML",
                        "name": "xml",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "default": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "summary": "Delete comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of deleting comment",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/posts/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get posts",
                "produces": [
                    "application/json"
                ],
                "summary": "List posts",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "posts filter by user",
                        "name": "userId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "show data like XML",
                        "name": "xml",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    },
                    "default": {
                        "description": ""
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "update post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update post",
                "parameters": [
                    {
                        "description": "JSON structure for updating post",
                        "name": "RequestPost",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.post"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "default": {
                        "description": ""
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "create post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create post",
                "parameters": [
                    {
                        "description": "JSON structure for creating post",
                        "name": "RequestPost",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.createPostStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "201": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get post by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Show a posts",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "show data like XML",
                        "name": "xml",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    },
                    "default": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "delete post by ID",
                "summary": "Delete post",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of deleting post",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/posts/{id}/comments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List comments like request /comments?postId={id}",
                "summary": "List comments of post",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of post",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "show data like XML",
                        "name": "xml",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "default": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "main.comment": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "postId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "main.createCommentStruct": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "postId": {
                    "type": "integer"
                }
            }
        },
        "main.createPostStruct": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "main.post": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "APIKey",
            "in": "header"
        }
    }
}